BACKEND: feat: Register (completed), update model 22.10.2022

+++ BackendPBL4/SyncFolderPBL4/src/main/java/com/SyncFolderPBL4/model/service/GenericService.java
@@ -1,7 +1,7 @@
 package com.SyncFolderPBL4.model.service;
 
 public interface GenericService<T> {
-	T findOne(Long id);
+	T findOne(Integer id);
 	T findOne(T obj);
-	Long save(T obj);
+	Integer save(T obj);
 }
